Fixed image loading in UIButton.cs
Images weren't loaded for UIButtons when a button was inside a
SetSharedProperty node in a UI script, and the the SetSharedProperty
node didn't set the image.

Also added event handlers for clicking on the SkinLight, SkinMedium and
SkinDark buttons.

Changed behaviour of HeadSkinBrowser and BodySkinBrowser to load all the
different appearances for the various skin tones.