More changes to UITextEdit
All the "standard" keys should now be mapped to input, except for "ø",
on Norwegian keyboards, "ö" on Swedish ones, and "æ" on Norwegian ones,
"ä" on Swedish ones, as well as "å" on Norwegian ones.
Backspace still doesn't work properly.

Found out that the Update method for UITextEdit was being called twice,
so disabled this.