More work on UITextEdit
Fixed an assignment bug which set m_VerticalTextBoundary to only the
last character in the single line text edit.

Bruteforced the cursor to only delete from the correct position. It's a
hack (the real solution is to figure out why the cursor is ahead or
behind the actual text), but it sorta works for now.